cmake: option for native optimized build
For native builds that are to be used on the host that
does the build, optimizize by passing -march=native,
this selects the specific march of the processor (e.g.
-march=broadwell or -march=core2 or -march=cortex-a8, etc)
and also selects relevant processor features (e.g. AES,
hard floating point, precise -mfpu like vfpv3-d16, etc.).

This also covers the AES feature (-maes). For example,
32-bit Core2 Duo does not have this feature, so with
-march=native, the compiler does not set -maes. On
Core i5, it does. On ARM GCC doesn't even have
-maes, so it won't get set.

Set NATIVE=OFF for builds that are ment to be used on other machines.
Then, no -march should be passed at all, which tells the compiler to
produce a generic binary.

This is an improvement over the "middle ground" of setting
-march=${ARCH}.

This holds for GCC and Clang. Except, for ARM, Clang (at least the Arch
Linux build) does include hard floating point feature but does not
include detailed FPU features even when -march=native is set. If we
really want to go beyond the default and optimize ARM binaries with
Clang, that should be done by adding a Platform/Linux-clang-armv7l.cmake
file, checking for processor features, and specifying flags.